feat(cli): add mode-toggle as installable UI component
Promote the docs site theme switcher to a first-class registry entry so
`solidui-cli add mode-toggle` pulls button and dropdown-menu deps and
installs light/dark/system switching without copy-paste from examples.

Uses lucide-solid icons (already credited in NOTICE). Update dark-mode
docs to reference the single add command.

Maintainers: run `pnpm build:registry` in apps/docs before release.
